query updates running better
